Move Player/Enemy-related types into player.rs/enemy.rs
